Enhanced Time/Date Display with Additional Line & Advanced Features (#174)
* Enhanced Time/Date Display with Additional Line & Advanced Features

An additional line has been included in the code to optimize the display of the time and date labels. The recommended approach is to display the time on the top row and the date on the bottom row. Furthermore, the day of the month can now be appended with the appropriate suffix (i.e., "st", "nd", "rd", or "th").